# apps/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Owostack API",
    "description": "Unified billing, entitlements, and usage metering API. Owostack sits between your app and payment providers (Paystack, Stripe, Dodo, Polar) to handle subscriptions, feature gating, metered usage, credit systems, and invoicing.",
    "version": "1.0.0",
    "contact": {
      "name": "Owostack",
      "url": "https://owostack.com"
    }
  },
  "servers": [
    {
      "url": "https://api.owostack.com",
      "description": "Production"
    },
    {
      "url": "https://sandbox.owostack.com",
      "description": "Sandbox"
    }
  ],
  "security": [{ "bearerAuth": [] }],
...
  "paths": {
...
    "/v1/billing/usage": {
      "get": {
        "operationId": "getBillingUsage",
        "tags": ["Billing"],
        "summary": "Get unbilled overage usage",
        "description": "Get unbilled overage usage for a customer with estimated costs per feature.",
        "parameters": [
          {
            "name": "customer",
            "in": "query",
            "required": true,
            "schema": { "type": "string" },
            "description": "Customer ID or email"
          }
        ],
        "responses": {
          "200": {
            "description": "Unbilled usage",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "customer": { "type": "string" },
                    "currency": { "type": "string" },
                    "totalEstimated": { "type": "number" },
                    "features": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "featureSlug": { "type": "string" },
                          "featureName": { "type": "string" },
                          "overageUnits": { "type": "number" },
                          "overagePrice": { "type": "number" },
                          "estimatedCost": { "type": "number" }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/v1/billing/invoice": {
      "post": {
        "operationId": "createInvoice",
        "tags": ["Billing"],
        "summary": "Generate an invoice",
        "description": "Generate an invoice for unbilled overage usage.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["customer"],
                "properties": {
                  "customer": { "type": "string" }
                }
              },
              "example": { "customer": "user_123" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Invoice generated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "invoice": { "$ref": "#/components/schemas/Invoice" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/v1/billing/invoices": {
      "get": {
        "operationId": "listInvoices",
        "tags": ["Billing"],
        "summary": "List invoices",
        "description": "List all invoices for a customer.",
        "parameters": [
          {
            "name": "customer",
            "in": "query",
            "required": true,
            "schema": { "type": "string" },
            "description": "Customer ID or email"
          }
        ],
        "responses": {
          "200": {
            "description": "Invoice list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "invoices": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/InvoiceSummary" }
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/v1/billing/invoice/{id}/pay": {
      "post": {
        "operationId": "payInvoice",
        "tags": ["Billing"],
        "summary": "Pay an invoice",
        "description": "Pay an invoice via auto-charge (if card on file) or checkout URL fallback.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "Invoice ID"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "callbackUrl": { "type": "string", "format": "uri" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Invoice payment result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "paid": { "type": "boolean" },
                    "checkoutUrl": { "type": "string", "format": "uri" },
                    "invoice": { "$ref": "#/components/schemas/InvoiceSummary" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key in the format `owo_sk_...`"
      }
    },
    "schemas": {
...
      "Invoice": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "number": { "type": "string" },
          "status": { "type": "string", "enum": ["open", "paid", "void"] },
          "total": { "type": "number" },
          "currency": { "type": "string" },
          "lineItems": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "featureSlug": { "type": "string" },
                "quantity": { "type": "number" },
                "unitPrice": { "type": "number" },
                "amount": { "type": "number" }
              }
            }
          }
        }
      },
      "InvoiceSummary": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "number": { "type": "string" },
          "status": { "type": "string" },
          "total": { "type": "number" },
          "currency": { "type": "string" },
          "createdAt": { "type": "number" },
          "paidAt": { "type": "number", "nullable": true }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "enum": [false] },
          "error": { "type": "string" }
        }
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Authentication failed",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": { "success": false, "error": "Invalid API Key" }
          }
        }
      },
      "BadRequest": {
        "description": "Invalid request",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": { "success": false, "error": "Invalid request body" }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": { "success": false, "error": "Not found" }
          }
        }
      }
    }
  }
}